pyproject.toml
[project]
name = "capstone-project-fastapi"
version = "0.1.0"
description = "FastAPI MVC Project with MYSQL for Capstone Project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiomysql>=0.3.2",
    "alembic>=1.18.0",
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]
